Add files via upload
The task:

Install the Apache Web server in the virtual machine. On the server, host the website you created in the first task.

Create a CGI-based Perl program that takes the data provided in the form, checks it and returns a response in HTML format depending on the theme you have chosen. For example, if the website is for scientific data processing, the Perl program should process the data (or simulate the processing if it is too complex) and return a generated web page with the results. The program must be secure, use "tainted" mode, and detect and report incomplete or invalid data. It is important that, in any case, specific information describing the problems encountered is returned to the user.

The CGI application must also be able to be accessed by the curl tool from the command line. Write brief documentation (in HTML format, accessible from the website) on what fields and what information needs to be fed to the CGI program in order to achieve the same thing that would be done in the form.