Add files via upload
Create a web page with embedded JavaScript code. When the web page loads, the JavaScript code in the web page must find the first 100 composite (not prime) numbers and report them in alert().The numbers in the message must be separated by a comma and a space, i.e. the start of the message must look like "4, 6, 8, 9".Use "use strict". Use only simple arithmetic functions to find prime numbers. Use only code you have written. During billing, you may be asked to change the number of prime numbers you are outputting, e.g. to increase to 200 or similar.