Add files via upload
Create a web page with a form that allows the user to upload an unlimited number of files organized in a file tree. It is not necessary to create a processing CGI script.
The form starts with the name of a directory into which the user can specify to upload as many files as he wants. The user must also be able to create other directories within the directory and to upload files and create directories within them as well. There shall be no limit to the number and depth of files and directories.

The form shall preserve the organization of the files and directories, i.e. it shall be possible to reconstruct a tree of files and directories on the server side from the data passed by the form, with the content uploaded by the user. The web page shall not reload when the user is filling in the form. The form must look neat, i.e. it must be possible to see which files belong to which directory, etc. Use "use strict" and only code you have written.

An example of a form in progress is shown in the image "tree.png" attached to the assignment.