Add files via upload
Create a web page with a text field and a button (not in a form). When the user enters a file name without an extension, JavaScript must look for files with the extensions ".html" and ".txt" in the same directory on the server as the web page itself.
In the case of one of these files (with the ".html" extension preferred), JavaScript must load its contents into an HTML <div> element. If none of these files exist, the text input field should turn red.

Use AJAX to search for files. Both files must be searched at the same time, i.e., asynchronously (successive await assignments are not asynchronous).When the user starts editing the text field, its color must reset.

The web page must not reload during user actions, use only JavaScript code. Use "use strict" and only code you write.