Add files via upload
Using Node.js, write a server-side application that downloads the specified HTML file (if it exists in the working directory) on user request. If there is no such file, the application should return an HTTP 404 error. Use only the 'http', 'url', 'querystring' and 'fs' modules. Use only asynchronous functions of the 'fs' module.